[create] 時間間隔指定でアイテムのログを取得 その他にルートをコントローラーにファイル分け
